Refactor: created a components folder
