+++ src/app/shared/components/card/card.component.scss
@@ -16,13 +16,12 @@
   .title {
     padding-top: 8px;
     padding-bottom: 8px;
-    color: #FFFFFF;
     font-size: 20px;
     max-height: 72px;
     overflow: hidden;
     white-space: nowrap;
     text-overflow: ellipsis;
-    color: pink;
+    color: #f48ea0;
   }
   .info {
     display: flex;
