+++ src/app/shared/components/grid/grid.component.scss
@@ -1,5 +1,4 @@
 .grid {
-  background-color: #081C24;
   display: grid;
   align-items: center;
   grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
