+++ src/app/shared/index.ts
@@ -1,1 +1,1 @@
-export * from './header/header.component';
+export * from './components/header/header.component';
