+++ src/app/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {CardComponent} from './card/card.component';
-import {GridComponent} from './grid/grid.component';
+import {CardComponent} from './components/card/card.component';
+import {GridComponent} from './components/grid/grid.component';
 import {BackdropImagePipe} from './pipes/backdrop-image.pipe';
 import {PosterImagePipe} from './pipes/poster-image.pipe';
 import {FontAwesomeTestingModule} from "@fortawesome/angular-fontawesome/testing";
